Implement USER CRUD, POST CRUD

+++ src/main/java/com/kjs990114/goodong/domain/user/Contribution.java
@@ -22,6 +22,6 @@ public class Contribution {
     @ManyToOne
     private User user;
 
-    private Short count = 1;
+    private Integer count = 1;
 
 }
